refactor: move secret-specific code out of from output.text.messages
